unified/api/requirements.txt
gunicorn==25.0.3
thursday==0.0
flasgger==0.9.4
apispec==2.0.2
gevent==25.9.1
gevent-websocket==0.10.1
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.3
python-magic==0.4.18
PyJWT==1.7.1
PythonVideoConverter==1.0.3
geopy==1.20.0
fpdf==1.7.2
PyPDF2==1.26.0
Faker==2.0.3
pycountry==19.8.18
coverage==4.5.4
tzlocal==1.5.1
ujson==1.35
pillow==8.4.0
pyfcm==2.0.7
langchain==1.2.10
langchain-openai==1.1.9
openai==2.20.0
ollama==0.6.1
langchain-ollama==1.0.1
